feat : API 연동 아크패시브 노드 UI추가
작업 완료
API에서 가져온 아크패시브 노드를 Profile UI에 추가하였습니다.

작업 예정
노드 각각 클릭가능, 상세 정보 확인 기능 추가
