+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/Common.kt
@@ -36,7 +36,7 @@ fun titleBoldWhite12() = TextStyle(
 @Composable
 fun Title(
     title: String,
-    onClick: () -> Unit,
+    onClick: () -> Unit = {},
 ) {
     Column(
         modifier = Modifier
