--- app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/arkPassive/ArkPassive.kt
@@ -1,168 +1,0 @@
-package com.hongmyeoun.goldcalc.view.profile.content.arkPassive
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Divider
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.bumptech.glide.integration.compose.placeholder
-import com.hongmyeoun.goldcalc.R
-import com.hongmyeoun.goldcalc.model.profile.arkpassive.ArkPassiveNode
-import com.hongmyeoun.goldcalc.ui.theme.ArkPassiveEnlightenment
-import com.hongmyeoun.goldcalc.ui.theme.ArkPassiveEvolution
-import com.hongmyeoun.goldcalc.ui.theme.ArkPassiveLeap
-import com.hongmyeoun.goldcalc.ui.theme.GoldCalcTheme
-import com.hongmyeoun.goldcalc.ui.theme.LightGrayTransBG
-import com.hongmyeoun.goldcalc.view.common.noRippleClickable
-
-@Preview(showBackground = true)
-@Composable
-fun ArkPassiveTest() {
-    GoldCalcTheme {
-        Column(modifier = Modifier
-            .background(LightGrayTransBG, RoundedCornerShape(8.dp))
-            .clip(RoundedCornerShape(8.dp))
-            .noRippleClickable { }
-            .padding(8.dp)
-        ) {
-            ArkPassiveNodeBoxTest("진화")
-            ArkPassiveNodeBoxTest("깨달음")
-            ArkPassiveNodeBoxTest("도약")
-        }
-    }
-}
-
-@Composable
-private fun ArkPassiveNodeBoxTest(
-    type: String
-) {
-    val arkPassiveIcon =
-        if (type == "진화") R.drawable.ico_arkpassive_evolution else if (type == "깨달음") R.drawable.ico_arkpassive_enlightenment else if (type == "도약") R.drawable.ico_arkpassive_leap else R.drawable.ico_arkpassive_evolution
-    val arkPassiveColor =
-        if (type == "진화") ArkPassiveEvolution else if (type == "깨달음") ArkPassiveEnlightenment else if (type == "도약") ArkPassiveLeap else Color.White
-
-    Column {
-        ArkPassiveNodeTableTest(arkPassiveColor, arkPassiveIcon)
-        Box(modifier = Modifier.padding(12.dp)) {
-            ArkPassiveContentsTest(arkPassiveColor)
-        }
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun ArkPassiveNodeTableTest(arkPassiveColor: Color, arkPassiveIcon: Int) {
-    Row(
-        horizontalArrangement = Arrangement.Center
-    ) {
-        Box(
-            modifier = Modifier
-                .border(
-                    width = 0.5.dp,
-                    color = arkPassiveColor,
-                    shape = RoundedCornerShape(16.dp),
-                )
-                .padding(4.dp)
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.Center
-            ) {
-                GlideImage(
-                    model = arkPassiveIcon,
-                    loading = placeholder(arkPassiveIcon),
-                    contentDescription = ""
-                )
-                Text(text = "진화 ", color = Color.White)
-                Text(text = "140", color = arkPassiveColor)
-                Spacer(modifier = Modifier.width(4.dp))
-            }
-        }
-        Divider(
-            modifier = Modifier.align(Alignment.CenterVertically),
-            thickness = 0.5.dp,
-            color = arkPassiveColor,
-        )
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun ArkPassiveContents(
-    arkPassiveNode: ArkPassiveNode
-) {
-    Row(
-        horizontalArrangement = Arrangement.Center
-    ) {
-        GlideImage(
-            modifier = Modifier.clip(RoundedCornerShape(16.dp)),
-            model = R.drawable.ark_passive_evolution_1,
-            loading = placeholder(R.drawable.ark_passive_evolution_1),
-            contentDescription = "아크패시브"
-        )
-        GlideImage(
-            modifier = Modifier.clip(RoundedCornerShape(16.dp)), model = arkPassiveNode.icon, contentDescription = "아크패시브"
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Text(
-            text = "1티어", color = Color.White
-        )
-        Text(
-            text = arkPassiveNode.tier, color = Color.White
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Text(
-            text = "치명 Lv.10",
-            color = ArkPassiveEvolution,
-        )
-        Text(
-            text = "${arkPassiveNode.name} ${arkPassiveNode.level}",
-            color = ArkPassiveEvolution,
-        )
-    }
-}
-
-@Composable
-@OptIn(ExperimentalGlideComposeApi::class)
-private fun ArkPassiveContentsTest(
-    textColor: Color,
-) {
-    Row(
-        horizontalArrangement = Arrangement.Center
-    ) {
-        GlideImage(
-            modifier = Modifier.clip(RoundedCornerShape(16.dp)),
-            model = R.drawable.ark_passive_evolution_1,
-            loading = placeholder(R.drawable.ark_passive_evolution_1),
-            contentDescription = "아크패시브"
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Text(
-            text = "1티어", color = Color.White
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Text(
-            text = "치명 Lv.10",
-            color = textColor,
-        )
-    }
-}
