+++ app/src/main/java/com/hongmyeoun/goldcalc/view/profile/content/skill/SkillView.kt
@@ -76,6 +76,7 @@ fun SkillView(
             Simple(skills, gemList, viewModel)
         }
     }
+    Spacer(modifier = Modifier.height(8.dp))
 }
 
 @Composable
